+++ examples/plot_basic_histogram.py
@@ -15,18 +15,28 @@ measurement = np.concatenate((
     rng.exponential(size=N_BKG),
     rng.normal(loc=2, scale=0.2, size=N_SIG),
 ))
-expectation_hist = Histogram(expectation, label="MC", histtype="stepfilled", alpha=1)
-measurement_hist = Histogram(measurement, label="dummy data")
+expectation_hist = Histogram(
+    expectation,
+    bins=50,
+    bins_range=(1.1, 4),
+    norm=False,
+    label="MC",
+    histtype="stepfilled",
+    alpha=1,
+)
+measurement_hist = Histogram(
+    measurement,
+    bins=50,
+    bins_range=(1.1, 4),
+    norm=False,
+    label="dummy data",
+)
 
 # Initialise histogram plot
 plot_histo = HistogramPlot(
     ylabel="Number of events",
     xlabel="Invariant mass $m$ [a.u.]",
     logy=False,
-    # bins=np.linspace(0, 5, 60),  # you can force a binning for the plot here
-    bins=50,  # you can also define an integer number for the number of bins
-    bins_range=(1.1, 4),  # only considered if bins is an integer
-    norm=False,
     atlas_first_tag="Simulation Internal",
     atlas_second_tag="Example for more general plot",
     figsize=(6, 5),