+++ examples/plot_data_mc.py
@@ -14,21 +14,45 @@ data = rng.normal(size=30_000)
 
 # Set up the real plot
 data_mc_plot = HistogramPlot(
-    bins_range=[-2, 2],
     n_ratio_panels=1,
     stacked=True,
-    norm=False,
 )
 
 # Set the plot title
 data_mc_plot.title = "Test Data/MC Plot"
 
 # Add the different MC contributions to the plot
-data_mc_plot.add(Histogram(mc1, label="MC Process 1"))
-data_mc_plot.add(Histogram(mc2, label="MC Process 2"))
+data_mc_plot.add(
+    Histogram(
+        mc1,
+        bins=40,
+        bins_range=[-2, 2],
+        label="MC Process 1",
+        norm=False,
+    )
+)
+data_mc_plot.add(
+    Histogram(
+        mc2,
+        bins=40,
+        bins_range=[-2, 2],
+        label="MC Process 2",
+        norm=False,
+    )
+)
 
 # Add the data
-data_mc_plot.add(Histogram(data, label="Data", is_data=True, colour="k"))
+data_mc_plot.add(
+    Histogram(
+        data,
+        bins=40,
+        bins_range=[-2, 2],
+        label="Data",
+        is_data=True,
+        colour="k",
+        norm=False,
+    )
+)
 
 # Draw the plot
 data_mc_plot.draw()