+++ examples/plot_rocs.py
@@ -53,6 +53,46 @@ rnnip_cjets_rej = calculate_rejection(discs_rnnip[is_b], discs_rnnip[is_c], sig_
 dips_ujets_rej = calculate_rejection(discs_dips[is_b], discs_dips[is_light], sig_eff)
 dips_cjets_rej = calculate_rejection(discs_dips[is_b], discs_dips[is_c], sig_eff)
 
+# Define now the actual ROC curve objects
+rnnip_ujets_roc = Roc(
+    sig_eff=sig_eff,
+    bkg_rej=rnnip_ujets_rej,
+    n_test=n_jets_light,
+    rej_class="ujets",
+    signal_class="bjets",
+    label="RNNIP",
+)
+dips_ujets_roc = Roc(
+    sig_eff=sig_eff,
+    bkg_rej=dips_ujets_rej,
+    n_test=n_jets_light,
+    rej_class="ujets",
+    signal_class="bjets",
+    label="DIPS r22",
+)
+rnnip_cjets_roc = Roc(
+    sig_eff=sig_eff,
+    bkg_rej=rnnip_cjets_rej,
+    n_test=n_jets_c,
+    rej_class="cjets",
+    signal_class="bjets",
+    label="RNNIP",
+)
+dips_cjets_roc = Roc(
+    sig_eff=sig_eff,
+    bkg_rej=dips_cjets_rej,
+    n_test=n_jets_c,
+    rej_class="cjets",
+    signal_class="bjets",
+    label="DIPS r22",
+)
+
+# ROC curve objects can also be stored as yaml or json files
+rnnip_ujets_roc.save("rnnip_ujets_roc.yaml")
+dips_ujets_roc.save("dips_ujets_roc.yaml")
+rnnip_cjets_roc.save("rnnip_cjets_roc.yaml")
+dips_cjets_roc.save("dips_cjets_roc.yaml")
+
 # here the plotting of the roc starts
 logger.info("Plotting ROC curves.")
 plot_roc = RocPlot(
@@ -63,51 +103,45 @@ plot_roc = RocPlot(
     figsize=(6.5, 6),
     y_scale=1.4,
 )
-plot_roc.add_roc(
-    Roc(
-        sig_eff,
-        rnnip_ujets_rej,
-        n_test=n_jets_light,
-        rej_class="ujets",
-        signal_class="bjets",
-        label="RNNIP",
-    ),
-    reference=True,
-)
-plot_roc.add_roc(
-    Roc(
-        sig_eff,
-        dips_ujets_rej,
-        n_test=n_jets_light,
-        rej_class="ujets",
-        signal_class="bjets",
-        label="DIPS r22",
-    ),
-)
-plot_roc.add_roc(
-    Roc(
-        sig_eff,
-        rnnip_cjets_rej,
-        n_test=n_jets_c,
-        rej_class="cjets",
-        signal_class="bjets",
-        label="RNNIP",
-    ),
-    reference=True,
-)
-plot_roc.add_roc(
-    Roc(
-        sig_eff,
-        dips_cjets_rej,
-        n_test=n_jets_c,
-        rej_class="cjets",
-        signal_class="bjets",
-        label="DIPS r22",
-    ),
-)
+
+# Add the ROC curve objects to the plot
+plot_roc.add_roc(roc_curve=rnnip_ujets_roc, reference=True)
+plot_roc.add_roc(roc_curve=dips_ujets_roc)
+plot_roc.add_roc(roc_curve=rnnip_cjets_roc, reference=True)
+plot_roc.add_roc(roc_curve=dips_cjets_roc)
+
 # setting which flavour rejection ratio is drawn in which ratio panel
 plot_roc.set_ratio_class(1, "ujets")
 plot_roc.set_ratio_class(2, "cjets")
 
 plot_roc.draw()
 plot_roc.savefig("roc.png", transparent=False)
+
+# If you want to load now the ROC curves from file, you can do so
+loaded_rnnip_ujets_roc = Roc.load("rnnip_ujets_roc.yaml")
+loaded_dips_ujets_roc = Roc.load("dips_ujets_roc.yaml")
+loaded_rnnip_cjets_roc = Roc.load("rnnip_cjets_roc.yaml", colour="red")
+loaded_dips_cjets_roc = Roc.load("dips_cjets_roc.yaml", colour="red")
+
+
+# Now init a new plot
+loaded_plot_roc = RocPlot(
+    n_ratio_panels=2,
+    ylabel="Background rejection",
+    xlabel="$b$-jet efficiency",
+    atlas_second_tag="$\\sqrt{s}=13$ TeV, Loaded ROC Curves",
+    figsize=(6.5, 6),
+    y_scale=1.4,
+)
+
+loaded_plot_roc.add_roc(roc_curve=rnnip_ujets_roc, reference=True)
+loaded_plot_roc.add_roc(roc_curve=dips_ujets_roc)
+loaded_plot_roc.add_roc(roc_curve=rnnip_cjets_roc, reference=True)
+loaded_plot_roc.add_roc(roc_curve=dips_cjets_roc)
+
+# setting which flavour rejection ratio is drawn in which ratio panel
+loaded_plot_roc.set_ratio_class(1, "ujets")
+loaded_plot_roc.set_ratio_class(2, "cjets")
+
+loaded_plot_roc.draw()
+loaded_plot_roc.savefig("roc_loaded.png", transparent=False)